[HELPS-981] Move React code into separate folder
